added composer to the php Dockerfile and remove vendor directory on upgrades and reinstall it via composer
On the upgrade from 22.0.01 to 23.0.02 an error was thrown, that a ContainerInterface in the vendor directory was not found.
The error was reproducable after setting back the database and web directory.
The solution from https://stackoverflow.com/a/59472021 worked, so I added composer to the build and automated removal and reinstallation of the vendor directory.